+++ confiture-web-app/src/router.ts
@@ -379,7 +379,6 @@ const router = createRouter({
     }
 
     if (to.hash) {
-      console.log("bleteintr");
       return getScrollToHashBehaviour(to.hash);
     }
 
@@ -442,7 +441,6 @@ router.afterEach(async (to, from) => {
 
     if (!to.hash && !isTabNavigation(to, from)) {
       document.body.setAttribute("tabindex", "-1");
-      console.log("isterntr");
       document.body.focus();
       document.body.removeAttribute("tabindex");
     }
@@ -487,7 +485,7 @@ function scrollToTop() {
  * Scrolls to a given hash (without leading '#').
  * There must be an HTML element with `id=hash` in the current page.
  */
-export function getScrollToHashBehaviour(hash: string): ScrollBehaviorResult {
+function getScrollToHashBehaviour(hash: string): ScrollBehaviorResult {
   return new Promise((resolve) => {
     const { stop } = useResizeObserver(document.body, () => {
       const hashEl = document.querySelector(
@@ -505,20 +503,6 @@ export function getScrollToHashBehaviour(hash: string): ScrollBehaviorResult {
       }
     });
   });
-}
-
-export function scrollToHash(hash: string) {
-  const hashEl = document.querySelector(
-    "#" + CSS.escape(hash.substring(1))
-  ) as HTMLElement;
-  if (hashEl) {
-    // force the focus on the element ta make the focus order make sense when user starts tabbing
-    const initalTabIndex = hashEl.getAttribute("tabindex");
-    hashEl.setAttribute("tabindex", "-1");
-    hashEl.focus();
-    initalTabIndex ? hashEl.setAttribute("tabindex", initalTabIndex) : hashEl.removeAttribute("tabindex");
-    hashEl.scrollIntoView();
-  }
 }
 
 function scrollToPosition(
